pyproject.toml
[tool.poetry]
name = "fluent-issues"
version = "0.1.0"
description = ""
authors = ["Tudor Popa"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
numpy = "^1.26.4"
jupyter = "^1.0.0"
pandas = "^2.2.2"
matplotlib = "^3.8.4"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
